Tile refill and fall animation

+++ Ovis/src/Platform/OpenGL/OpenGLRendererAPI.h
@@ -11,7 +11,7 @@ namespace Ovis
 	public:
 		void Init() override;
 		void SetViewport(int width, int height) override;
-		void SetClearColor(glm::uvec4 color) override;
+		void SetClearColor(glm::vec4 color) override;
 		void Clear() override;
 
 		void DrawIndexed(const std::shared_ptr<VertexArray>& vertexArray, uint32_t indexCount = 0) override;
